Splitting Job Service from ULR
